Remove default when checking evo model name
Removed defaulting to GTR or WAG when requesting model type from string, should be handled
later in the code.
Fix tests that check default behaviour.
Moved tests for model types from substitution models to evolutionary models where the enums
are defined.
